+++ migrations/20250801122050_create_subcategory_table.sql
@@ -2,9 +2,9 @@
 -- +goose StatementBegin
 SELECT 'up SQL query';
 CREATE TABLE subcategory(
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(100) UNIQUE NOT NULL,
-    category_id INT NOT NULL REFERENCES category(id) ON DELETE CASCADE,
+    category_id BIGINT NOT NULL REFERENCES category(id) ON DELETE CASCADE,
     created_at TIMESTAMP NOT NULL DEFAULT now(),
     updated_at TIMESTAMP
 );
